Fixed Homepage Redirect Preventing Subsequent Homepage Login
Same issue as journey page, but in this instance user had to first do an incorrect login attempt.

In both cases, the issue was that '/index.php' would display at the end of the url pathname after redirect. Fix for both cases was to redirect to the directory of the filename so only '/' would display at the end of the url pathname (since file named index.php).